DateRange {CohortIncidence}	R Documentation

R6 Class Representing a DataRange

Description

The DateRange class, encapsulating the startDate and endDate of a date range.

Details

This class is used to specify a DateRange, with start and end dates specified as strings formatted as YYYY-MM-DD.

Active bindings

startDate

a character with format YYYY-MM-DD to be used as the date range's start date.

endDate

a character with format YYYY-MM-DD to be used as the date range's end date.

Methods

Public methods

DateRange$new()

DateRange$toList()

DateRange$asJSON()

DateRange$clone()

Method new()

creates a new instance, using the provided data param if provided.

Usage

DateRange$new(data = list())

Arguments

data

the data (as a json string or list) to initialize with

Method toList()

returns the R6 class elements as a list for use in jsonlite::toJSON()

Usage

DateRange$toList()

Method asJSON()

returns the JSON string for this R6 class

Usage

DateRange$asJSON()

Method clone()

The objects of this class are cloneable with this method.

Usage

DateRange$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package CohortIncidence version 3.1.2 Index]