IncidenceDesign {CohortIncidence}	R Documentation

R6 Class Representing a IncidenceDesign

Description

This class encapsulates the other R6 Class elements that define an IncidenceDesign

Details

The IncidenceDesign class encapsulates the following: - Cohort Definitions - Target Definitions - Outcome Definitions - Time At Risk Definitions - A List of Analyses - Concept Sets - Subgruops - Strata Settings Note, when serializing with a library such as jsonlite, first call toList() on the R6 class before calling jsonlite::toJSON(), or call toJSON directy on this class.

Active bindings

cohortDefs

A list of cohort definitions. Must be a list of class "CohortDefinition"

conceptSets

A list of concept set expressions. Currently unused.

targetDefs

A list of cohort references to be used as target cohorts. Must be a list of class "CohortReference"

outcomeDefs

A list of cohort references to be used as target cohorts. Must be a list of class "CohortReference"

timeAtRiskDefs

A list of cohort references to be used as target cohorts. Must be a list of class "CohortReference"

analysisList

A list of cohort references to be used as target cohorts. Must be a list of class "CohortReference"

subgroups

A list of cohort references to be used as target cohorts. Must be a list of class "CohortReference"

strataSettings

The strata settings for this design. Must be a class "StrataSettings"

studyWindow

a study window for this design. Must be a list of class "DateRange"

Methods

Public methods

IncidenceDesign$new()

IncidenceDesign$toList()

IncidenceDesign$asJSON()

IncidenceDesign$clone()

Method new()

creates a new instance, using the provided data param if provided.

Usage

IncidenceDesign$new(data = list())

Arguments

data

the data (as a json string or list) to initialize with

Method toList()

returns the R6 class elements as a list for use in jsonlite::toJSON()

Usage

IncidenceDesign$toList()

Method asJSON()

returns the JSON string for this R6 class

Usage

IncidenceDesign$asJSON(...)

Arguments

...

paramaters that are passed forward to rjsonlite::toJSON()

Method clone()

The objects of this class are cloneable with this method.

Usage

IncidenceDesign$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

[Package CohortIncidence version 3.1.2 Index]