createIncidenceDesign {CohortIncidence}	R Documentation

Creates R6 object for IncidenceDesign

Description

Creates R6 object for IncidenceDesign

Usage

createIncidenceDesign(
  cohortDefs,
  targetDefs,
  outcomeDefs,
  tars,
  analysisList,
  conceptSets,
  subgroups,
  strataSettings,
  studyWindow
)

Arguments

cohortDefs
The set of cohort definitions. Optional.

targetDefs
A list of target definitions, each element must be class CohortReference.

outcomeDefs
A list of outcome definitions, each element must be class Outcome

tars
A list of TAR definitions, each element must be class TimeAtRisk

analysisList
A list of analysis definitions, each element must be class IncidenceAnalysis

conceptSets
A list of concept sets, currently unused.

subgroups
A list of cohort subgroups, each element must be class Subgroup.

strataSettings
The strata settings used in the anlaysis, must be class StrataSettings.

studyWindow
Limits time at risk to the specified study window. Must be class DateRange.

Value

a R6 class: IncidenceDesign.

[Package CohortIncidence version 3.1.2 Index]