createStrataSettings {CohortIncidence}	R Documentation

Creates R6 object for StrataSettings

Description

Creates R6 object for StrataSettings

Usage

createStrataSettings(byAge = F, byGender = F, byYear = F, ageBreaks)

Arguments

byAge
a boolean indicating to stratify by age, defaults to F

byGender
a boolean indicating to stratify by gender, defaults to F

byYear
a boolean indicating to stratify by year, defaults to F

ageBreaks
a vector of integers indicating the age group bounds.

Value

an R list containing name-value pairs that will serialize into a org.ohdsi.analysis.cohortincidence.design.StratifySettings JSON format.

[Package CohortIncidence version 3.1.2 Index]